integrated_mine_platform/requirements.txt
# 集成矿山智能预测平台 - Python依赖

# Django核心
Django==5.0.0
djangorestframework==3.14.0
django-cors-headers==4.3.0

# 数据库
psycopg2-binary==2.9.9

# Redis 缓存
redis==5.0.1
django-redis==5.4.0

# 生产环境 Web 服务器
gunicorn==21.2.0

# 环境变量管理
python-dotenv==1.0.0

# 异步任务队列
celery==5.3.4

# 数据处理
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
joblib==1.3.2

# 深度学习
torch==2.1.0
torchvision==0.16.0
einops==0.7.0

# 机器学习库
lightgbm==4.1.0
xgboost==2.0.3
pytorch-forecasting==1.0.0

# 数据可视化
matplotlib==3.7.2
seaborn==0.12.2

# 小波分析
PyWavelets==1.4.1

# 科学计算
scipy==1.11.2

# 工具库
python-dateutil==2.8.2
pytz==2023.3
pillow==10.0.0
pyyaml==6.0.1

# 开发工具
autopep8==2.0.4
